python/safe_wallet/requirements.txt
py-clob-client-v2
web3
python-dotenv
requests
eth-abi
